requirements.txt
Flask==2.0.1
requests==2.26.0
oauthlib==3.1.1
pyOpenSSL==20.0.1
Flask-login==0.5.0
python-dotenv
Flask-PyMongo